fix: Fixes issue with cached static gump strings. Fixes bad gump colors (#1771)
### Summary
- Fixes an issue that causes CUO to crash due to bad string caching in static gumps
- Fixes wrong/bad 16bit html gump hues.
- Moves `C16232` (16-bit to 32-bit) and `C32216` (32-bit to 16-bit) to Utility class for broader use.

TODO:
- Some gumps have different 32bit (for string content) vs 16bit (for localized content) strings. Does this matter?

+++ Projects/UOContent/Gumps/StaticWarningGump.cs
@@ -42,7 +42,7 @@ public abstract class StaticWarningGump<T> : StaticGump<T> where T : StaticWarni
 
         builder.AddImageTiled(10, 10, width - 20, 20, 2624);
         builder.AddAlphaRegion(10, 10, width - 20, 20);
-        builder.AddHtmlLocalized(10, 10, width - 20, 20, header, (short)headerColor);
+        builder.AddHtmlLocalized(10, 10, width - 20, 20, header, headerColor);
 
         builder.AddImageTiled(10, 40, width - 20, height - 80, 2624);
         builder.AddAlphaRegion(10, 40, width - 20, height - 80);
